функции отображения перенесены во View
