+++ js/game.js
@@ -1,31 +1,26 @@
-import {getElementFromTemplate, changePageTemplate, addHandlerBackGreeting} from './utils.js';
+import {changePageTemplate} from './utils.js';
 import {getData, getState, setState} from './data';
-import getGameTemplate from './gameView';
+import {getGameModule} from './gameView';
 import stats from './stats';
 
-const levels = getData().levels;
 const stateModule = {question1: ``, question2: ``};
 
 function getModule() {
+  const levels = getData().levels;
   const curLevel = getState().curLevel;
   if (curLevel < levels.length) {
-    const level = levels[curLevel];
-    const template = getGameTemplate(level, getState());
-    const curModuleGame = getElementFromTemplate(template);
-    addEventHandlers(curModuleGame, level);
-    imgProportion(curModuleGame);
-    return addHandlerBackGreeting(curModuleGame);
+    return getGameModule(levels[curLevel], getState());
   } else {
     return stats;
   }
 }
 
-function nextModule() {
+export function nextModule() {
   setState(`curLevel`, getState().curLevel + 1);
   return getModule();
 }
 
-function checkComplete(e) {
+export function checkComplete(e) {
   stateModule[e.currentTarget.name] = e.currentTarget.value;
   let complete = true;
   let key;
@@ -39,53 +34,4 @@ function checkComplete(e) {
   }
 }
 
-function addEventHandlers(curModuleGame, level) {
-  let items;
-  if (level.type === `twoPicture`) {
-    items = Array.from(curModuleGame.querySelectorAll(`.game__option input`));
-    items.forEach((item) => {
-      item.addEventListener(`change`, (event) => {
-        checkComplete(event);
-      });
-    });
-  } else if (level.type === `onePicture`) {
-    items = Array.from(curModuleGame.querySelectorAll(`.game__option input`));
-    items.forEach((item) => {
-      item.addEventListener(`click`, (event) => {
-        changePageTemplate(nextModule());
-      });
-    });
-  } else if (level.type === `threePicture`) {
-    items = Array.from(curModuleGame.querySelectorAll(`.game__option`));
-    items.forEach((item) => {
-      item.addEventListener(`mousedown`, (event) => {
-        if (event.which === 1) {
-          changePageTemplate(nextModule());
-        }
-      });
-    });
-  }
-}
-
-function imgProportion(curModuleGame) {
-  const imgs = Array.from(curModuleGame.querySelectorAll(`.game__option img`));
-  imgs.forEach((item) => {
-    item.addEventListener(`load`, (event) => {
-      setProportions(item);
-    });
-  });
-}
-
-function setProportions(item) {
-  const width = parseInt(getComputedStyle(item.parentNode, ``).width, 10);
-  const height = parseInt(getComputedStyle(item.parentNode, ``).height, 10);
-  if (item.naturalWidth / item.naturalHeight > width / height) {
-    item.width = width;
-  } else {
-    item.height = height;
-  }
-}
-
-let moduleGame = getModule();
-
-export default moduleGame;
+export default getModule();
